Move ValidateApplicationService to a separate package for reuse
