+++ clientapi/routing/register_test.go
@@ -22,7 +22,6 @@ import (
 	"net/http"
 	"net/http/httptest"
 	"reflect"
-	"regexp"
 	"strings"
 	"testing"
 	"time"
@@ -32,7 +31,6 @@ import (
 	"github.com/matrix-org/dendrite/internal"
 	"github.com/matrix-org/dendrite/keyserver"
 	"github.com/matrix-org/dendrite/roomserver"
-	"github.com/matrix-org/dendrite/setup/config"
 	"github.com/matrix-org/dendrite/test"
 	"github.com/matrix-org/dendrite/test/testrig"
 	"github.com/matrix-org/dendrite/userapi"
@@ -165,70 +163,6 @@ func TestEmptyCompletedFlows(t *testing.T) {
 	// check for []
 	if ret == nil || len(ret) != 0 {
 		t.Error("Empty Completed Flow Stages should be a empty slice: returned ", ret, ". Should be []")
-	}
-}
-
-// This method tests validation of the provided Application Service token and
-// username that they're registering
-func TestValidationOfApplicationServices(t *testing.T) {
-	// Set up application service namespaces
-	regex := "@_appservice_.*"
-	regexp, err := regexp.Compile(regex)
-	if err != nil {
-		t.Errorf("Error compiling regex: %s", regex)
-	}
-
-	fakeNamespace := config.ApplicationServiceNamespace{
-		Exclusive:    true,
-		Regex:        regex,
-		RegexpObject: regexp,
-	}
-
-	// Create a fake application service
-	fakeID := "FakeAS"
-	fakeSenderLocalpart := "_appservice_bot"
-	fakeApplicationService := config.ApplicationService{
-		ID:              fakeID,
-		URL:             "null",
-		ASToken:         "1234",
-		HSToken:         "4321",
-		SenderLocalpart: fakeSenderLocalpart,
-		NamespaceMap: map[string][]config.ApplicationServiceNamespace{
-			"users": {fakeNamespace},
-		},
-	}
-
-	// Set up a config
-	fakeConfig := &config.Dendrite{}
-	fakeConfig.Defaults(config.DefaultOpts{
-		Generate:   true,
-		Monolithic: true,
-	})
-	fakeConfig.Global.ServerName = "localhost"
-	fakeConfig.ClientAPI.Derived.ApplicationServices = []config.ApplicationService{fakeApplicationService}
-
-	// Access token is correct, user_id omitted so we are acting as SenderLocalpart
-	asID, resp := validateApplicationService(&fakeConfig.ClientAPI, fakeSenderLocalpart, "1234")
-	if resp != nil || asID != fakeID {
-		t.Errorf("appservice should have validated and returned correct ID: %s", resp.JSON)
-	}
-
-	// Access token is incorrect, user_id omitted so we are acting as SenderLocalpart
-	asID, resp = validateApplicationService(&fakeConfig.ClientAPI, fakeSenderLocalpart, "xxxx")
-	if resp == nil || asID == fakeID {
-		t.Errorf("access_token should have been marked as invalid")
-	}
-
-	// Access token is correct, acting as valid user_id
-	asID, resp = validateApplicationService(&fakeConfig.ClientAPI, "_appservice_bob", "1234")
-	if resp != nil || asID != fakeID {
-		t.Errorf("access_token and user_id should've been valid: %s", resp.JSON)
-	}
-
-	// Access token is correct, acting as invalid user_id
-	asID, resp = validateApplicationService(&fakeConfig.ClientAPI, "_something_else", "1234")
-	if resp == nil || asID == fakeID {
-		t.Errorf("user_id should not have been valid: @_something_else:localhost")
 	}
 }
